BPUB-949 move WatchRecord class to separate file from litecoin-, bitcoin- & RPC- blockchain watcher
